+++ deploy/prod-root/prod-root-stop.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
-echo "Stop running docker compose"
-cd A2rchi-prod-root/deploy/prod-root/
-docker compose -f prod-root-compose.yaml down
+if [ -z "$(ls -A A2rchi-prod-root/deploy/prod-root/)" ]; then
+    echo "Deployment directory is empty; skipping docker compose down"
+else
+    echo "Stop running docker compose"
+    cd A2rchi-prod-root/deploy/prod-root/
+    docker compose -f prod-root-compose.yaml down
+fi
